Upgraded docs, removed unnecessary files
As I work towards a functional and visually appealing reverse literate
configuration I need to actively minimize my work such that it becomes
tenable to document this entire project. In order to do this, I did my
first round of clean up, and will continue to do them in the future.